+++ out/no_proxy.js
@@ -4,9 +4,9 @@ exports.applyNoProxyBypass = applyNoProxyBypass;
 exports.collectNoProxyHosts = collectNoProxyHosts;
 exports.getDirectConnectEnv = getDirectConnectEnv;
 const fs = require("fs");
+const os = require("os");
 const path = require("path");
-const os = require("os");const os = require("os");
-const path = require("path");const proxy_config_loader_1 = require("./proxy_config_loader");
+const proxy_config_loader_1 = require("./proxy_config_loader");
 const PROXY_ENV_KEYS = [
     "HTTP_PROXY",
     "HTTPS_PROXY",